Input: synaptics-rmi4 - setup input_dev parent
Fixes a pending FIXME to set input_dev->dev.parent correctly.
Since we'll soon be discarding the per-function input_devs and
using just a single input_dev for all the functions on a particular
rmi_dev, we use the rmi_dev->dev as the parent device for the
input_dev.

Signed-off-by: Christopher Heiny <cheiny@synaptics.com>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>