Input: psmouse - Add API for passthrough buttons
This allows drivers to redirect button presses from a different device
to a psmouse device. This is particularly useful when dealing with
situations such as the one we have right now with the Lenovo ThinkPad
T450s and X1 Carbon, where the physical trackpoint buttons are routed
through the touchpad as opposed to the TrackPoint, and as such need to
be redirected in the driver to go to the proper device.

Signed-off-by: Stephen Chandler Paul <thatslyude@gmail.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>