Input: synaptics-rmi4: Change the rmi device's name to be less generic
Currently, the rmi device has the name sensorXX. Sensor is very generic
and in cases when there is little context it is hard to tell what the
device is (ie in /sys/devices or /proc/interrupts). This patch changes
the name to rmiXX which at least indicates that it is an rmi device.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>