Input: synaptics-rmi4: Use the transport driver's input device
If the transport driver already has an input device allocated for it,
then allow syanptics-rmi4 driver to use it instead of creating a new input
device. One example is some HID touchpads with external buttons. The
fireware reports button clicks with pass through HID reports which are
not seen by the RMI driver. Reusing the input report allocated by
hid-input will allow those reports to be sent to the host.

Signed-off-by: Andrew Duggan <aduggan@synaptics.com>