hid-lenovo: Add support for Thinkpad X1 Tablet Cover
Add IDs for the Lenovo X1 Cover at hid-lenovo driver.
Let the hid device of usb if0 and if1 bind to hid-lenovo driver only
because usb if2 is a rmi device and has to be handled by the rmi driver.

This adds support for X1 function keys and LED control.

Signed-off-by: Dennis Wassenberg <dennis.wassenberg@secunet.com>