hid-lenovo: configure Lenovo X1 cover initial state
The thinkpad x1 cover behaves different in cases when it was connected
during boot and is was not conncted during boot

For example if it was connected during boot an activated FN lock led means
that FN is locked. If it was not connected during boot an activated FN lock led
means FN is not locked. Thats why we have to set a predefined initial state.

This patch will configure usb IF1 hid device of the Lenovo X1 Tablet cover
like the Windows driver will do this is case of activcated FN lock led
means FN is locked.

Signed-off-by: Dennis Wassenberg <dennis.wassenberg@secunet.com>