rmi: Introduced input_configured step in rmi flow
RMI F03 gets its data from both. From the transport device data
or from rmi_read_block. In case of probing the rmi f03 the data is passed by
the transport device. But these data did not arrive at the rmi f03 function handler
because they are rejected by hid-rmi because the RMI_STARTED flag was not set.
This flag will be set if probing of the functions and fuction handlers is finished.
That means we have to relocate the serio device registration to an input_configured step.
During the real probing phase the f03 will be initialized (no transport device data are
needed in this step). After that the transport device is successfully registered and
the RMI_STARTED flag will be set. After this the input is ready to be configured.
This is now done by input_configured. In case of F03 the serio port will be registered.

Signed-off-by: Dennis Wassenberg <dennis.wassenberg@secunet.com>