rmi: f03: make serio port registration reliable
Because serio port registration is made asynchronous it is possible that
hid-rmi has stopped device io (at the end of during input_configured phase).
If the serio device registration runs after stopping io the device
is not able to reply to any commands. This will lead to a timeout
and the serio device registration will fail.

Signed-off-by: Dennis Wassenberg <dennis.wassenberg@secunet.com>